# OtakuQuest/otakuquest.client/src/api/generated/models/AuthResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AuthResponseDto = {
    token: string | null;
    message: string | null;
};