# OtakuQuest/otakuquest.client/src/api/generated/models/CompleteTaskResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CompleteTaskResponseDto = {
    message: string | null;
    xpReward: number;
    currencyReward: number;
    strengthReward: number;
    intelligenceReward: number;
    defenceReward: number;
    newLevel: number;
    currentXP: number;
    xpToNextLevel: number;
};