# OtakuQuest/otakuquest.client/src/api/generated/models/PlayerStatsDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PlayerStatsDto = {
    username: string | null;
    level: number;
    xp: number;
    currency: number;
    str: number;
    int: number;
    def: number;
    currentHP: number;
    maxHP: number;
    avatarItemId: number | null;
};