# OtakuQuest/otakuquest.client/src/api/generated/models/TaskStatus.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum TaskStatus {
    '_0' = 0,
    '_1' = 1,
    '_2' = 2,
    '_3' = 3,
}